@@ -8,18 +8,16 @@ use App\Http\Resources\RecursosHumanos\NominaPrestamos\PrestamoEmpresarialResour
 use App\Models\RecursosHumanos\NominaPrestamos\PlazoPrestamoEmpresarial;
 use App\Models\RecursosHumanos\NominaPrestamos\PrestamoEmpresarial;
 use App\Models\User;
-use Carbon\Carbon;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Log;
 use Illuminate\Validation\ValidationException;
 use Src\Shared\Utils;
 
 class PrestamoEmpresarialController extends Controller
 {
-    private $entidad = 'Prestamo Empresarial';
+    private string $entidad = 'Prestamo Empresarial';
     public function __construct()
     {
         $this->middleware('can:puede.ver.prestamo_empresarial')->only('index', 'show');
@@ -28,41 +26,31 @@ class PrestamoEmpresarialController extends Controller
         $this->middleware('can:puede.eliminar.prestamo_empresarial')->only('destroy');
     }
 
-    public function index(Request $request)
+    public function index()
     {
-        $results = [];
-
-
         $usuario = Auth::user();
         $usuario_ac = User::where('id', $usuario->id)->first();
         if ($usuario_ac->hasRole([User::ROL_GERENTE, User::ROL_RECURSOS_HUMANOS, User::ROL_CONTABILIDAD, User::ROL_ADMINISTRADOR])) {
             $results = PrestamoEmpresarial::ignoreRequest(['campos'])->filter()->orderBy('id', 'desc')->get();
-            $results = PrestamoEmpresarialResource::collection($results);
-            return response()->json(compact('results'));
         } else {
-
             $results = PrestamoEmpresarial::where('solicitante', $usuario->empleado->id)->ignoreRequest(['campos'])->filter()->get();
-            $results = PrestamoEmpresarialResource::collection($results);
-            return response()->json(compact('results'));
         }
-
-
-
+        $results = PrestamoEmpresarialResource::collection($results);
         return response()->json(compact('results'));
     }
-    public function show(Request $request, PrestamoEmpresarial $prestamoEmpresarial)
+    public function show(PrestamoEmpresarial $prestamo)
     {
-        $modelo = new PrestamoEmpresarialResource($prestamoEmpresarial);
-        return response()->json(compact('modelo'), 200);
+        $modelo = new PrestamoEmpresarialResource($prestamo);
+        return response()->json(compact('modelo'));
     }
     public function store(PrestamoEmpresarialRequest $request)
     {
         try {
             DB::beginTransaction();
             $datos = $request->validated();
-            $prestamoEmpresarial = PrestamoEmpresarial::create($datos);
-            $this->crear_plazos($prestamoEmpresarial, $request->plazos);
-            $modelo = new PrestamoEmpresarialResource($prestamoEmpresarial);
+            $prestamo = PrestamoEmpresarial::create($datos);
+            $this->crearPlazos($prestamo, $request->plazos);
+            $modelo = new PrestamoEmpresarialResource($prestamo);
             $mensaje = Utils::obtenerMensaje($this->entidad, 'store');
             DB::commit();
             return response()->json(compact('mensaje', 'modelo'));
@@ -73,39 +61,38 @@ class PrestamoEmpresarialController extends Controller
             ]);
         }
     }
-    public function update(PrestamoEmpresarialRequest $request, PrestamoEmpresarial $prestamoEmpresarial)
+    public function update(PrestamoEmpresarialRequest $request, PrestamoEmpresarial $prestamo)
     {
         $datos = $request->validated();
-        $prestamoEmpresarial->update($datos);
-        $this->modificar_plazo($prestamoEmpresarial, $request->plazos);
-        $modelo = new PrestamoEmpresarialResource($prestamoEmpresarial);
+        $prestamo->update($datos);
+        $this->modificarPlazo($request->plazos);
+        $modelo = new PrestamoEmpresarialResource($prestamo);
         $mensaje = Utils::obtenerMensaje($this->entidad, 'update');
         return response()->json(compact('mensaje', 'modelo'));
     }
-    public function destroy(Request $request, PrestamoEmpresarial $prestamoEmpresarial)
+    public function destroy(PrestamoEmpresarial $prestamo)
     {
-        $deleted = PlazoPrestamoEmpresarial::where('id_prestamo_empresarial', $prestamoEmpresarial->id)->delete();
-        $prestamoEmpresarial->delete();
-        $modelo = $prestamoEmpresarial;
+        $prestamo->delete();
+        $modelo = $prestamo;
         $mensaje = Utils::obtenerMensaje($this->entidad, 'destroy');
         return response()->json(compact('mensaje', 'modelo'));
     }
     public function deshabilitarPrestamo(Request $request)
     {
-        $prestamoEmpresarial = PrestamoEmpresarial::where('id',$request->id)->first();
-        $prestamoEmpresarial->motivo= $request->motivo;
-        $prestamoEmpresarial->estado= 'INACTIVO';
-        $prestamoEmpresarial->save();
-        $prestamoEmpresarial->plazo_prestamo_empresarial_info()->update(['estado' => false]);
-        $modelo = $prestamoEmpresarial;
+        $prestamo_empresarial = PrestamoEmpresarial::where('id',$request->id)->first();
+        $prestamo_empresarial->motivo= $request->motivo;
+        $prestamo_empresarial->estado= 'INACTIVO';
+        $prestamo_empresarial->save();
+        $prestamo_empresarial->plazo_prestamo_empresarial_info()->update(['estado' => false]);
+        $modelo = $prestamo_empresarial;
         $mensaje = Utils::obtenerMensaje($this->entidad, 'destroy');
         return response()->json(compact('mensaje', 'modelo'));
     }
-    public function crear_plazos(PrestamoEmpresarial $prestamoEmpresarial, $plazos)
+    public function crearPlazos(PrestamoEmpresarial $prestamo, $plazos)
     {
-        $plazosActualizados = collect($plazos)->map(function ($plazo) use ($prestamoEmpresarial) {
+        $plazos_actualizados = collect($plazos)->map(function ($plazo) use ($prestamo) {
             return [
-                'id_prestamo_empresarial' => $prestamoEmpresarial->id,
+                'id_prestamo_empresarial' => $prestamo->id,
                 'pago_couta' => false,
                 'num_cuota' => $plazo['num_cuota'],
                 'fecha_vencimiento' =>  $plazo['fecha_vencimiento'],
@@ -114,12 +101,12 @@ class PrestamoEmpresarialController extends Controller
 
             ];
         })->toArray();
-        PlazoPrestamoEmpresarial::insert($plazosActualizados);
+        PlazoPrestamoEmpresarial::insert($plazos_actualizados);
     }
-    public function modificar_plazo(PrestamoEmpresarial $prestamoEmpresarial, $plazos)
+    public function modificarPlazo($plazos)
     {
         // Crear un arreglo con los datos actualizados para cada plazo
-        $plazosActualizados = collect($plazos)->map(function ($plazo) {
+        $plazos_actualizados = collect($plazos)->map(function ($plazo) {
             return [
                 'id' =>  $plazo['id'],
                 'pago_couta' =>  $plazo['pago_couta'],
@@ -130,12 +117,12 @@ class PrestamoEmpresarialController extends Controller
             ];
         })->toArray();
         // Realizar la actualización masiva
-        foreach ($plazosActualizados as $plazoActualizado) {
-            $id = $plazoActualizado['id'];
-            PlazoPrestamoEmpresarial::where('id', $id)->update($plazoActualizado);
+        foreach ($plazos_actualizados as $plazo) {
+            $id = $plazo['id'];
+            PlazoPrestamoEmpresarial::where('id', $id)->update($plazo);
         }
     }
-    public function obtener_prestamo_empleado(Request $request)
+    public function obtenerPrestamoEmpleado(Request $request)
     {
             list($mes, $anio) = explode('-',$request->mes);
             $results = PlazoPrestamoEmpresarial::
