@@ -28,7 +28,7 @@ class VacacionRequest extends FormRequest
         return [
             'empleado_id' => 'required|exists:empleados,id',
             'periodo_id' => 'required|exists:periodos,id',
-            'derecho_vacaciones' => 'nullable|date_format:Y-m-d',
+            'derecho_vacaciones' => 'nullable',
             'fecha_inicio' => 'nullable|date_format:Y-m-d',
             'fecha_fin' => 'nullable|date_format:Y-m-d',
             'fecha_inicio_rango1_vacaciones' => 'nullable|date_format:Y-m-d',
@@ -53,7 +53,6 @@ class VacacionRequest extends FormRequest
             'fecha_fin_rango1_vacaciones',
             'fecha_inicio_rango2_vacaciones',
             'fecha_fin_rango2_vacaciones',
-            'derecho_vacaciones'
         ];
         foreach ($dateFields as $field) {
             if ($this->$field) {
