@@ -28,10 +28,10 @@ class PermisoEmpleadoResource extends JsonResource
             'estado' => $this->estado_permiso_id,
             'estado_permiso_info' => $this->estadoPermiso?->nombre,
             'empleado' => $this->empleado_id,
-            'empleado_info' => $this->empleadoInfo?->nombres . ' ' . $this->empleadoInfo?->apellidos,
-            'departamento' => $this->empleadoInfo?->departamento->nombre,
-            'id_jefe_inmediato' => $this->empleadoInfo?->jefe->id,
-            'jefe_inmediato' => $this->empleadoInfo?->jefe->nombres . '' . $this->empleadoInfo?->jefe->apellidos,
+            'empleado_info' => $this->empleado?->nombres . ' ' . $this->empleado?->apellidos,
+            'departamento' => $this->empleado?->departamento->nombre,
+            'id_jefe_inmediato' => $this->empleado?->jefe->id,
+            'jefe_inmediato' => $this->empleado?->jefe->nombres . '' . $this->empleado?->jefe->apellidos,
             'fecha_hora_solicitud' => Carbon::parse($this->created_at)->format($this->mask),
             'fecha_hora_reagendamiento' => $this->fecha_hora_reagendamiento ? Carbon::parse($this->fecha_hora_reagendamiento)->format($this->mask) : null,
             'nombre' => $this->documento != null ? json_decode($this->documento)->nombre : '',
