@@ -2,7 +2,7 @@
 
 namespace App\Http\Resources\RecursosHumanos\NominaPrestamos;
 
-use Carbon\Carbon;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
 class PrestamoEmpresarialResource extends JsonResource
@@ -10,17 +10,16 @@ class PrestamoEmpresarialResource extends JsonResource
     /**
      * Transform the resource into an array.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
+     * @param  Request  $request
+     * @return array
      */
     public function toArray($request)
     {
-        $controller_method = $request->route()->getActionMethod();
-        $modelo = [
+        return [
             'id' => $this->id,
             'solicitante' => $this->solicitante,
             'solicitante_info' => $this->empleado_info->nombres . ' ' . $this->empleado_info->apellidos,
-            'fecha' =>  $this->cambiar_fecha($this->fecha),
+            'fecha' => $this->fecha,
             'monto' =>  $this->monto,
             'periodo' =>   $this->periodo_id,
             'periodo_info' => $this->periodo_info? $this->periodo_info->nombre:'' ,
@@ -31,11 +30,6 @@ class PrestamoEmpresarialResource extends JsonResource
 
 
         ];
-        return $modelo;
     }
-    private function cambiar_fecha($fecha)
-    {
-        $fecha_formateada = Carbon::parse($fecha)->format('d-m-Y');
-        return $fecha_formateada;
-    }
+
 }
