@@ -7,12 +7,13 @@ use Illuminate\Http\Request;
 abstract class ReporteCuestionarioAbstract
 {
     public const CODIGOS_ALCOHOL_DROGAS = [
-        4,
-        4.1,
-        4.2,
-        4.3,
-        4.4,
-        4.5,
+        4,  // 0 - Principal droga que consume
+        4.1, // 1 - En caso de seleccionar otros especfique cual
+        4.2, // 2 - Frecuencia de consum
+        4.3, // 3 - Empleado reconoce eneer n problema de consumo
+        4.4, // 4 - Factores psicosociles relaconados al consumo
+        4.5, // 5 - En caso de seleccionar otros especifique cual
+        4.6, // 6 - Desea recibir tratamiento en caso de ser consumidor
     ];
 
     // Excel
